test: extract MockConfigRepository
