+++ internal/command/application/usecases/addcommand_test.go
@@ -11,13 +11,14 @@ import (
 	commanddomain "gomander/internal/command/domain"
 	"gomander/internal/command/domain/test"
 	configdomain "gomander/internal/config/domain"
+	test2 "gomander/internal/config/domain/test"
 )
 
 func TestDefaultAddCommand_Execute(t *testing.T) {
 	t.Run("Should add the command", func(t *testing.T) {
 		// Arrange
 		mockCommandRepository := new(test.MockCommandRepository)
-		mockUserConfigRepository := new(MockConfigRepository)
+		mockUserConfigRepository := new(test2.MockConfigRepository)
 
 		projectId := "project1"
 		sut := usecases.NewAddCommand(mockUserConfigRepository, mockCommandRepository)
@@ -55,7 +56,7 @@ func TestDefaultAddCommand_Execute(t *testing.T) {
 	t.Run("Should return an error if fails to get the user config", func(t *testing.T) {
 		// Arrange
 		mockCommandRepository := new(test.MockCommandRepository)
-		mockUserConfigRepository := new(MockConfigRepository)
+		mockUserConfigRepository := new(test2.MockConfigRepository)
 
 		sut := usecases.NewAddCommand(mockUserConfigRepository, mockCommandRepository)
 		newCommandBuilder := test.NewCommandBuilder()
@@ -77,7 +78,7 @@ func TestDefaultAddCommand_Execute(t *testing.T) {
 	t.Run("Should return an error if fails to get all commands", func(t *testing.T) {
 		// Arrange
 		mockCommandRepository := new(test.MockCommandRepository)
-		mockUserConfigRepository := new(MockConfigRepository)
+		mockUserConfigRepository := new(test2.MockConfigRepository)
 
 		projectId := "project1"
 		sut := usecases.NewAddCommand(mockUserConfigRepository, mockCommandRepository)
@@ -106,7 +107,7 @@ func TestDefaultAddCommand_Execute(t *testing.T) {
 	t.Run("Should return an error if fails to create commands", func(t *testing.T) {
 		// Arrange
 		mockCommandRepository := new(test.MockCommandRepository)
-		mockUserConfigRepository := new(MockConfigRepository)
+		mockUserConfigRepository := new(test2.MockConfigRepository)
 
 		projectId := "project1"
 		sut := usecases.NewAddCommand(mockUserConfigRepository, mockCommandRepository)
