+++ internal/command/application/usecases/reordercommands_test.go
@@ -11,13 +11,14 @@ import (
 	commanddomain "gomander/internal/command/domain"
 	"gomander/internal/command/domain/test"
 	configdomain "gomander/internal/config/domain"
+	test2 "gomander/internal/config/domain/test"
 )
 
 func TestDefaultReorderCommands_Execute(t *testing.T) {
 	t.Run("Should reorder commands", func(t *testing.T) {
 		// Arrange
 		mockCommandRepository := new(test.MockCommandRepository)
-		mockUserConfigRepository := new(MockConfigRepository)
+		mockUserConfigRepository := new(test2.MockConfigRepository)
 
 		projectId := "project1"
 		sut := usecases.NewReorderCommands(mockUserConfigRepository, mockCommandRepository)
@@ -60,7 +61,7 @@ func TestDefaultReorderCommands_Execute(t *testing.T) {
 	t.Run("Should return an error if fails to get the user config", func(t *testing.T) {
 		// Arrange
 		mockCommandRepository := new(test.MockCommandRepository)
-		mockUserConfigRepository := new(MockConfigRepository)
+		mockUserConfigRepository := new(test2.MockConfigRepository)
 
 		sut := usecases.NewReorderCommands(mockUserConfigRepository, mockCommandRepository)
 		expectedErr := errors.New("failed to get user config")
@@ -80,7 +81,7 @@ func TestDefaultReorderCommands_Execute(t *testing.T) {
 	t.Run("Should return an error if fails to retrieve commands", func(t *testing.T) {
 		// Arrange
 		mockCommandRepository := new(test.MockCommandRepository)
-		mockUserConfigRepository := new(MockConfigRepository)
+		mockUserConfigRepository := new(test2.MockConfigRepository)
 
 		projectId := "project1"
 		sut := usecases.NewReorderCommands(mockUserConfigRepository, mockCommandRepository)
@@ -105,7 +106,7 @@ func TestDefaultReorderCommands_Execute(t *testing.T) {
 	t.Run("Should return an error if fails to update commands", func(t *testing.T) {
 		// Arrange
 		mockCommandRepository := new(test.MockCommandRepository)
-		mockUserConfigRepository := new(MockConfigRepository)
+		mockUserConfigRepository := new(test2.MockConfigRepository)
 
 		projectId := "project1"
 		sut := usecases.NewReorderCommands(mockUserConfigRepository, mockCommandRepository)
