+++ internal/config/application/usecases/shared_test.go
@@ -1,24 +1,1 @@
 package usecases_test
-
-import (
-	"github.com/stretchr/testify/mock"
-
-	"gomander/internal/config/domain"
-)
-
-type MockUserConfigRepository struct {
-	mock.Mock
-}
-
-func (m *MockUserConfigRepository) GetOrCreate() (*domain.Config, error) {
-	args := m.Called()
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*domain.Config), args.Error(1)
-}
-
-func (m *MockUserConfigRepository) Update(config *domain.Config) error {
-	args := m.Called(config)
-	return args.Error(0)
-}
