+++ internal/project/application/usecases/openproject_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stretchr/testify/mock"
 
 	"gomander/internal/config/domain"
+	test2 "gomander/internal/config/domain/test"
 	"gomander/internal/project/application/usecases"
 	projectdomain "gomander/internal/project/domain"
 	"gomander/internal/project/domain/test"
@@ -16,7 +17,7 @@ import (
 func TestDefaultOpenProject_Execute(t *testing.T) {
 	t.Run("Should open a project successfully", func(t *testing.T) {
 		// Arrange
-		mockConfigRepository := new(MockConfigRepository)
+		mockConfigRepository := new(test2.MockConfigRepository)
 		mockProjectRepository := new(test.MockProjectRepository)
 
 		sut := usecases.NewOpenProject(mockConfigRepository, mockProjectRepository)
@@ -72,7 +73,7 @@ func TestDefaultOpenProject_Execute(t *testing.T) {
 
 	t.Run("Should return an error if getting the config fails", func(t *testing.T) {
 		// Arrange
-		mockConfigRepository := new(MockConfigRepository)
+		mockConfigRepository := new(test2.MockConfigRepository)
 		mockProjectRepository := new(test.MockProjectRepository)
 
 		sut := usecases.NewOpenProject(mockConfigRepository, mockProjectRepository)
@@ -91,7 +92,7 @@ func TestDefaultOpenProject_Execute(t *testing.T) {
 
 	t.Run("Should return an error if updating the config fails", func(t *testing.T) {
 		// Arrange
-		mockConfigRepository := new(MockConfigRepository)
+		mockConfigRepository := new(test2.MockConfigRepository)
 		mockProjectRepository := new(test.MockProjectRepository)
 
 		sut := usecases.NewOpenProject(mockConfigRepository, mockProjectRepository)
